Split ClrAppDomain into its own source file
Move ClrAppDomain into its own source file.
